+++ R/class_builder.R
@@ -228,12 +228,10 @@ target_debug.tar_builder <- function(target) {
   if (should_debug) {
     # Covered in tests/interactive/test-debug.R
     # nocov start
-    target$command$expr <- as.expression(
-      list(
-        instructions = quote(targets::tar_debug_instructions()),
-        browser = quote(browser()),
-        expr = target$command$expr
-      )
+    target$command$expr <- c(
+      expression(targets::tar_debug_instructions()),
+      expression(browser()),
+      target$command$expr
     )
     target$cue$mode <- "always"
     target$settings$deployment <- "main"